@@ -1,6 +1,8 @@
+/* eslint-disable react-refresh/only-export-components */
+import { Metadata } from "next";
+
 import Breadcrumb from "@/components/Common/Breadcrumb";
 import NotFound from "@/components/NotFound";
-import { Metadata } from "next";
 
 export const metadata: Metadata = {
   title: "404 Page | Play SaaS Starter Kit and Boilerplate for Next.js",
