@@ -1,12 +1,13 @@
-import React from "react";
 import "./styles.css";
+
+import React from "react";
 import { useTranslation } from "react-i18next";
 
 const BackgroundVideo = () => {
   const { t } = useTranslation()
   return (
     <div className="video-container">
-        
+
       <video
         className="bg-video"
         autoPlay
@@ -21,7 +22,7 @@ const BackgroundVideo = () => {
         <div
           className="hero-content wow fadeInUp mx-auto max-w-[760px] flex items-end justify-center text-center h-[79vh]"
           data-wow-delay=".2s"
-          >
+        >
           <h1 className="text-3xl font-bold leading-snug text-white sm:text-4xl sm:leading-snug lg:text-4xl lg:leading-[1.2] max-[700px]:text-3xl">
             {t("hero.title")}
           </h1>
